feat: Implement new date picker component
Adds a new date picker component to the application. This component allows the user to select a date, and the selected date is then used to filter the data that is displayed.

This new component is used on the location page to allow the user to select a date to view the prayer times for.